Update.gitignore to upload some files
Re-adds custom-redacted.css to give an example of how to add a tiling
background to your server, and our updating list of ipbans.txt to show
which IPs we've banned on our server. In the future we'll comment the
entries (if possible) so that you know the name of the person who got
banned, their probable alternate account if any, and the reason for it.