Added Mixedmons, a custom metagame of ours
In addition, I've also set some of the lesser played metagames to
searchShow: false in order to hide them from the 'Look for a Battle'
list.

Mixedmons is a metagame that I came up with by combining two popular
metagames, the current OM of the Month, STABmons, and Ability Exchange.
With both of these combined, it's quite possible to make very unexpected
teams. Your team can have whatever move you want it to have under the
condition that it's the same type as your Pokemon, and it can even have
any ability as long as another member of your team has it.

The banlist for this metagame is currently shown here, but the metagame
will be fully documented on www.pokesun.com relatively soon.